project/backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
cryptography==42.0.5